+++ spyne/test/interface/test_xml_schema.py
@@ -17,6 +17,7 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
 #
 
+
 import logging
 import unittest
 
@@ -39,7 +40,9 @@ from spyne.model.primitive import Uuid
 from spyne.protocol.soap import Soap11
 from spyne.service import ServiceBase
 from spyne.util.xml import get_schema_documents
-from spyne.util.xml import parse_schema
+from spyne.util.xml import parse_schema_element
+from spyne.util.xml import parse_schema_string
+
 from spyne.interface.xml_schema import XmlSchema
 
 
@@ -149,7 +152,8 @@ class TestXmlSchema(unittest.TestCase):
             out_protocol=Soap11()
         )
 
-        _ns = {'xs': "http://www.w3.org/2001/XMLSchema"}
+        _ns = {'xs': ns.xsd}
+        pref_xs = ns.const_prefmap[ns.xsd]
         xs = XmlSchema(app.interface)
         xs.build_interface_document()
         elt = xs.get_interface_document()['tns'].xpath(
@@ -157,11 +161,11 @@ class TestXmlSchema(unittest.TestCase):
                     namespaces=_ns)[0]
 
         assert elt.xpath('//xs:element[@name="base64_1"]/@type',
-                                        namespaces=_ns)[0] == 'xs:base64Binary'
+                            namespaces=_ns)[0] == '%s:base64Binary' % pref_xs
         assert elt.xpath('//xs:element[@name="base64_2"]/@type',
-                                        namespaces=_ns)[0] == 'xs:base64Binary'
+                            namespaces=_ns)[0] == '%s:base64Binary' % pref_xs
         assert elt.xpath('//xs:element[@name="hex"]/@type',
-                                        namespaces=_ns)[0] == 'xs:hexBinary'
+                            namespaces=_ns)[0] == '%s:hexBinary' % pref_xs
 
 
     def test_multilevel_customized_simple_type(self):
@@ -198,7 +202,7 @@ class TestXmlSchema(unittest.TestCase):
         assert any[0].attrib['namespace'] == '##other'
         assert any[0].attrib['processContents'] == 'lax'
 
-    def test_xml_data(self):
+    def _test_xml_data(self):
         tns = 'kickass.ns'
         class ProductEdition(ComplexModel):
             __namespace__ = tns
@@ -233,7 +237,7 @@ class TestXmlSchema(unittest.TestCase):
                                     '/xs:sequence/xs:element[@name="edition"]'
                 '/xs:complexType/xs:simpleContent/xs:extension'
                                     '/xs:attribute[@name="id"]'
-                ,namespaces=app.interface.nsmap)) == 1
+                ,namespaces={'xs': ns.xsd})) == 1
 
     def test_subs(self):
         from lxml import etree
@@ -279,7 +283,7 @@ class TestParseOwnXmlSchema(unittest.TestCase):
         schema = get_schema_documents([SomeGuy], tns)['tns']
         print etree.tostring(schema, pretty_print=True)
 
-        objects = parse_schema(schema)
+        objects = parse_schema_element(schema)
         pprint(objects[tns].types)
 
         NewGuy = objects[tns].types["SomeGuy"]
@@ -297,7 +301,7 @@ class TestParseOwnXmlSchema(unittest.TestCase):
         schema = get_schema_documents([SomeGuy], tns)['tns']
         print etree.tostring(schema, pretty_print=True)
 
-        objects = parse_schema(schema)
+        objects = parse_schema_element(schema)
         pprint(objects[tns].types)
 
         NewGuy = objects['some_ns'].types["SomeGuy"]
@@ -315,7 +319,7 @@ class TestParseOwnXmlSchema(unittest.TestCase):
         schema = get_schema_documents([SomeGuy], tns)['tns']
         print etree.tostring(schema, pretty_print=True)
 
-        objects = parse_schema(schema)
+        objects = parse_schema_element(schema)
         pprint(objects[tns].types)
 
         NewGuy = objects['some_ns'].types["SomeGuy"]
@@ -338,7 +342,7 @@ class TestParseForeignXmlSchema(unittest.TestCase):
     </xsd:complexType>
 </xsd:schema>"""
 
-        objects = parse_schema(etree.fromstring(schema))
+        objects = parse_schema_string(schema)
         pprint(objects[tns].types)
 
         NewGuy = objects[tns].types['SomeGuy']